% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gliab.R
\name{calc_gliab}
\alias{calc_gliab}
\title{Calculation of genetic liabilities}
\usage{
calc_gliab(obj, beta, mu, sigma)
}
\arguments{
\item{obj}{A matrix consisting of 0s, 1s or 2s, where each row represents a genotype.}

\item{beta}{List of beta values.}

\item{mu}{List of mu values to normalize on.}

\item{sigma}{List of sigma values to normalize on.}
}
\value{
The function returns a vector of genetic liabilities for each genotype.
}
\description{
This is an internal helper function to used to calculate genetic liabilities.
}
\keyword{internal}
