% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_normalization_consts.R
\name{calc_normalization_consts}
\alias{calc_normalization_consts}
\title{Calculation of normalization constants}
\usage{
calc_normalization_consts(MAF, causal)
}
\arguments{
\item{MAF}{Vector of Minor Allele Frequencies.}

\item{causal}{Vector of causal SNPs where causal SNPs has the value 1 and non causal SNPs has the value 0.}
}
\value{
A list with that contains two vectors, one for theoretical values of mu and one for the theoretical values of
sigma.
}
\description{
This an internal helper function used to calculate the theoretical value of the constants
mu and sigma, which are used to normalize the data.
}
\keyword{internal}
