% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_calc.R
\name{cond_calc}
\alias{cond_calc}
\title{Conditional values of mu and sigma}
\usage{
cond_calc(i, covmatrix)
}
\arguments{
\item{i}{parameter index of interest.}

\item{covmatrix}{the co-variance matrix.}
}
\value{
The conditional values of mu and sigma.
}
\description{
This function is an internal helper function used to calculate the
conditional mu and sigma of a co-variance for a multivariate normal distribution.
}
\keyword{internal}
