% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmatrix.R
\name{covmatrix}
\alias{covmatrix}
\title{Calculate co-variance matrix for liability modeling}
\usage{
covmatrix(h2, n_sib = 0)
}
\arguments{
\item{h2}{Integer specifying heritability parameter (0 < h2 < 1).}

\item{sib}{Integer specifying number of siblings (default 0).}
}
\value{
A co-variance matrix, created from the value of h2 and
the number of sibs.
}
\description{
This function creates the theoretical co-variance matrix for the
multivariate normal distribution which is used to model the liabilities.
}
