% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRds.R
\name{createRds}
\alias{createRds}
\title{Create .rds file.}
\usage{
createRds(path, nrow, ncol)
}
\arguments{
\item{path}{Path where to save the files (DO NOT SPECIFIY FILE EXTENSION).}

\item{nrow}{Integer specifying number of rows.}

\item{ncol}{Integer specifying number of columns.}
}
\value{
A list containing a file-backed matrix, a
FAM tibble and a MAP tibble.
}
\description{
This function is used to create and save a file-backed matrix with related information.
}
\details{
When using this function both a .bk and .rds file will be created at the specified path. Besides
the file-backed matrix two tibbles containing FAM (phenotype information) and MAP (SNP information)
will be intialized and saved together with the file-backed matrix in the .rds file.
Both the .bk and the .rds file must always be in the same directory.
The function openRds can be used to access these files again (see function reference for openRds).
}
