% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get statistics on how many SNPs have been correctly identified}
\usage{
get_stats(pvalues, true_causal, a)
}
\arguments{
\item{pvalues}{A vector of the pvalues calculated from analysis.}

\item{true_causal}{A vector indicating causal SNPs (1 where causal, 0 where not causal).}

\item{a}{An integer specifying significance level and thereby the cutoff for a SNP being causal.}
}
\value{
a dataframe containing the statistics
}
\description{
This functions gives statistics on the count and frequency of false negatives, true negatives,
false positives and true positives in analysed data
}
