% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.R
\name{GWAS}
\alias{GWAS}
\title{Perform genome wide association study (GWAS)}
\usage{
GWAS(rds.obj, y, include = bigparallelr::rows_along(rds.obj$genotypes))
}
\arguments{
\item{rds.obj}{A list of length 3 containing an FBM.code256 named genotypes and accompanying FAM and MAP.}

\item{y}{Vector of regressands to regress on (case-control status of the genotypes in person)}

\item{include}{Vector of rows to use in regression. Used with cross-validation. Default uses all rows.}
}
\value{
A data.frame with slopes of each regression, standard errors of each slope, t-scores associated with each slope and P-values of each slope.
}
\description{
This function performs a GWAS on given genotype and phenotype data to find out how statistically associated SNPs are with a target disease.
}
