% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFH.R
\name{LTFH}
\alias{LTFH}
\title{Perform GWAS on posterior mean genetic liabilities (LTFH)}
\usage{
LTFH(rds.obj, prevalence, h2)
}
\arguments{
\item{rds.obj}{A list object with an FBM.code256 and accompanying FAM and MAP tibbles.}

\item{prevalence}{The likelihood of having the disease in the population.}

\item{h2}{The heritability parameter for the disease.}
}
\value{
A list with 3 values: the GWAS output, the calculated posterior mean genetic liabilities, and a data frame with the configurations and the associated posterior mean genetic liabilities.
}
\description{
This function finds all the unique status configurations in the data and
calculates the mean posterior genetic liability for each of these unique
configurations using gibbs sampling. The calculated posterior mean genetic
liabilities are then matched to each person in the dataset and used to perform GWAS.
}
