% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power_plots.R
\name{Power_plots}
\alias{Power_plots}
\title{Create power plots from GWAS, GWAX and LTFH data}
\usage{
Power_plots(Gwas_data, p)
}
\arguments{
\item{Gwas_data}{A list with exactly three entrances, with the GWAS-values gained from from each model (GWAS, GWAX and LTFH), where the data from GWAS has to be the first entrance, the data from Gwax has to be the second entrance and the data from LTFH has to be on the last entrance.}

\item{p}{significance level used to determine causal SNPs}
}
\value{
Ggplot with power plots from GWAS, GWAX and LTFH, to get a visual comparison of the prediction power for each model.
}
\description{
This function is used to calculate and plot the power for each of the models GWAS, GWAX and LTFH.
This function is only design to plot all three plots against each other to get a visual comparison of the power plots for each model.
}
