% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_status.R
\name{Predict_status}
\alias{Predict_status}
\title{Predict Status for new data with pre-existing model}
\usage{
Predict_status(rds.obj, model, configs = NULL, prevalence = 0.05)
}
\arguments{
\item{rds.obj}{A rds object containing an FBM.code256.}

\item{model}{A model to use for predicting.}

\item{configs}{A data frame with configurations and their liabilities. Only used for LTFH.}

\item{prevalence}{The likelihood of having the disease in the population.}
}
\value{
A vector of 1's and 0's indicating predicted status.
}
\description{
This function is used to predict the status of one or more new persons.
}
