% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRds.R
\name{saveRds}
\alias{saveRds}
\title{Saving the work done to an rds object}
\usage{
saveRds(rds.obj)
}
\arguments{
\item{rds.obj}{The rds object to save.}
}
\description{
This function is used to save the changes made to an rds object during an R session.
}
