% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_disease.R
\name{sim_disease}
\alias{sim_disease}
\title{Simulation of disease parameters}
\usage{
sim_disease(
  n_snp,
  prevalence,
  h2,
  causal = NULL,
  causal_n = max(1, round(n_snp/10)),
  maf = NULL,
  maf_low = 0.01,
  maf_high = 0.49
)
}
\arguments{
\item{n_snp}{Integer specifying amount of SNPs.}

\item{prevalence}{Integer specifying prevalence of disease in simulated population.}

\item{h2}{Integer specifying heritability parameter.}

\item{causal}{Vector of predetermined causal SNPs (0 if not causal, 1 if causal). Leave empty to get random causal SNPs.}

\item{causal_n}{Integer specifying amount of causal SNPs if no causal param given. Default value is 10 procent of total SNPs.}

\item{maf}{Vector of predetermined Minor Allele Frequencies.Leave empty to get random MAF.}

\item{maf_low}{Integer specifying lower bound for MAF if none given.}

\item{maf_high}{Integer specifying upper bound for MAF if none given.}
}
\value{
A list with all the disease parameters.
}
\description{
This function is used to simulate disease parameters. If not predetermined,
the function will randomly calculate causal SNPs, MAF vaules and beta values.
}
