% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFH.R
\name{LTFH}
\alias{LTFH}
\title{Perform GWAS on posterior mean genetic liabilities}
\usage{
LTFH(child, prevalence, h2)
}
\arguments{
\item{child}{A list object with an FBM.code256 and accompanying FAM and MAP.}

\item{prevalence}{The likelihood of having the disease in the population}

\item{h2}{Heritability parameter.}
}
\value{
The function returns a list, where the first entrance is the GWAS
output and the second entrance are the calculated posterior mean genetic liabilities.
}
\description{
This function finds all the unique status configurations in the data and
calculates the mean posterior genetic liability for each of these unique
configurations using gibbs sampling. The calculated posterior mean genetic
liabilities are then matched to each person in the dataset and used to perform GWAS.
}
