% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{Manhattan_plot}
\alias{Manhattan_plot}
\title{Make a Manhattan plot from analysis data}
\usage{
Manhattan_plot(x)
}
\arguments{
\item{List}{object containing gwas data from an analysis.}
}
\value{
A Manhattan plot with each regressor (SNP) on the x-axis and their -log10(P-values) on the y-axis.
}
\description{
This function produces a Manhattan plot from p-values found in an association analysis.
}
