% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{Manhattan_plot}
\alias{Manhattan_plot}
\title{Make a Manhattan plot from analysis data}
\usage{
Manhattan_plot(x, a)
}
\arguments{
\item{x}{List object containing vector with numbering on SNPs and vector with P-values. Example could be GWAS data from analysis.}

\item{a}{significance level used to determine whether SNPs are causal or not.}
}
\value{
A Manhattan plot with each regressor (SNP) on the x-axis and their -log10(P-values) on the y-axis.
}
\description{
This function produces a Manhattan plot from P-values found in an association analysis.
}
