% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_trunc.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{Sample from a truncated normal distribution}
\usage{
rnorm_trunc(n, range, mu, sigma)
}
\arguments{
\item{n}{Integer value specifying the number of samples.}

\item{range}{vector with two values specifying the truncated range (lower bound and upper bound).}

\item{mu}{Integer value specifying the mean of normal distribution.}

\item{sigma}{Integer value specifying the standard deviation of the normal distribution.}
}
\value{
Sample values from the truncated normal distribution.
}
\description{
Helper function used to sample from a truncated normal distribution.
}
